refactor pdlb to pdb, add can filtering, etc.
