+++ config/pdb.yaml
@@ -1,12 +1,12 @@
-struct_name: pdlb_config_t
+struct_name: pdb_config_t
 chip: STM32G431CBTx
 regs:
-- name: pdlb_can_id
+- name: can_id
   type: uint8
 - name: jetson_can_id
   type: uint8
 can_filtering:
-  id_reg: pdlb_can_id
+  id_reg: can_id
   id_type: ext
   delay_compensation: true
   tdc_offset: 13
