+++ src/pdb/pdb.ioc
@@ -129,8 +129,8 @@ ProjectManager.MainLocation=Src
 ProjectManager.NoMain=false
 ProjectManager.PreviousToolchain=
 ProjectManager.ProjectBuild=false
-ProjectManager.ProjectFileName=pdlb.ioc
-ProjectManager.ProjectName=pdlb
+ProjectManager.ProjectFileName=pdb.ioc
+ProjectManager.ProjectName=pdb
 ProjectManager.ProjectStructure=
 ProjectManager.RegisterCallBack=
 ProjectManager.StackSize=0x400
