+++ src/pdb/startup_stm32g431xx.s
