--- src/pdlb/Inc/main.h
@@ -1,81 +1,0 @@
-/* USER CODE BEGIN Header */
-/**
-  ******************************************************************************
-  * @file           : main.h
-  * @brief          : Header for main.c file.
-  *                   This file contains the common defines of the application.
-  ******************************************************************************
-  * @attention
-  *
-  * Copyright (c) 2026 STMicroelectronics.
-  * All rights reserved.
-  *
-  * This software is licensed under terms that can be found in the LICENSE file
-  * in the root directory of this software component.
-  * If no LICENSE file comes with this software, it is provided AS-IS.
-  *
-  ******************************************************************************
-  */
-/* USER CODE END Header */
-
-/* Define to prevent recursive inclusion -------------------------------------*/
-#ifndef __MAIN_H
-#define __MAIN_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/* Includes ------------------------------------------------------------------*/
-#include "stm32g4xx_hal.h"
-
-/* Private includes ----------------------------------------------------------*/
-/* USER CODE BEGIN Includes */
-
-/* USER CODE END Includes */
-
-/* Exported types ------------------------------------------------------------*/
-/* USER CODE BEGIN ET */
-
-/* USER CODE END ET */
-
-/* Exported constants --------------------------------------------------------*/
-/* USER CODE BEGIN EC */
-
-/* USER CODE END EC */
-
-/* Exported macro ------------------------------------------------------------*/
-/* USER CODE BEGIN EM */
-
-/* USER CODE END EM */
-
-/* Exported functions prototypes ---------------------------------------------*/
-void Error_Handler(void);
-
-/* USER CODE BEGIN EFP */
-void PostInit(void);
-/* USER CODE END EFP */
-
-/* Private defines -----------------------------------------------------------*/
-#define CAN_RX_LED_Pin GPIO_PIN_8
-#define CAN_RX_LED_GPIO_Port GPIOA
-#define CAN_TX_LED_Pin GPIO_PIN_9
-#define CAN_TX_LED_GPIO_Port GPIOA
-#define CAN_STB_Pin GPIO_PIN_10
-#define CAN_STB_GPIO_Port GPIOA
-#define AUTON_LED_B_Pin GPIO_PIN_5
-#define AUTON_LED_B_GPIO_Port GPIOB
-#define AUTON_LED_G_Pin GPIO_PIN_6
-#define AUTON_LED_G_GPIO_Port GPIOB
-#define AUTON_LED_R_Pin GPIO_PIN_7
-#define AUTON_LED_R_GPIO_Port GPIOB
-
-/* USER CODE BEGIN Private defines */
-
-/* USER CODE END Private defines */
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* __MAIN_H */
